fix: merge ProductType extensions on a single file
