+++ packages/smooth_app/lib/query/product_query.dart
@@ -1,7 +1,6 @@
 import 'dart:ui';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:openfoodfacts/openfoodfacts.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:smooth_app/data_models/preferences/user_preferences.dart';
@@ -9,6 +8,7 @@ import 'package:smooth_app/database/dao_string.dart';
 import 'package:smooth_app/database/local_database.dart';
 import 'package:smooth_app/helpers/analytics_helper.dart';
 import 'package:smooth_app/pages/preferences/user_preferences_dev_mode.dart';
+import 'package:smooth_app/pages/product/product_type_extensions.dart';
 import 'package:uuid/uuid.dart';
 
 // ignore: avoid_classes_with_only_static_members
@@ -294,50 +294,4 @@ abstract class ProductQuery {
         ProductField.OBSOLETE,
         ProductField.OWNER_FIELDS,
       ];
-}
-
-extension ProductTypeExtension on ProductType {
-  String getDomain() => switch (this) {
-        ProductType.food => 'openfoodfacts',
-        ProductType.beauty => 'openbeautyfacts',
-        ProductType.petFood => 'openpetfoodfacts',
-        ProductType.product => 'openproductsfacts',
-      };
-
-  String getLabel(final AppLocalizations appLocalizations) => switch (this) {
-        ProductType.food => appLocalizations.product_type_label_food,
-        ProductType.beauty => appLocalizations.product_type_label_beauty,
-        ProductType.petFood => appLocalizations.product_type_label_pet_food,
-        ProductType.product => appLocalizations.product_type_label_product,
-      };
-
-  String getRoadToScoreLabel(final AppLocalizations appLocalizations) =>
-      switch (this) {
-        ProductType.food => appLocalizations.hey_incomplete_product_message,
-        ProductType.beauty =>
-          appLocalizations.hey_incomplete_product_message_beauty,
-        ProductType.petFood =>
-          appLocalizations.hey_incomplete_product_message_pet_food,
-        ProductType.product =>
-          appLocalizations.hey_incomplete_product_message_product,
-      };
-
-  String getShareProductLabel(
-    final AppLocalizations appLocalizations,
-    final String url,
-  ) =>
-      switch (this) {
-        ProductType.food => appLocalizations.share_product_text(
-            url,
-          ),
-        ProductType.beauty => appLocalizations.share_product_text_beauty(
-            url,
-          ),
-        ProductType.petFood => appLocalizations.share_product_text_pet_food(
-            url,
-          ),
-        ProductType.product => appLocalizations.share_product_text_product(
-            url,
-          ),
-      };
 }
